Refactor Query options
