+++ src/module.ts
@@ -1,7 +1,8 @@
 import { ConfigEditor, QueryEditor } from 'components';
 import { DataSourcePlugin } from '@grafana/data';
 import { DataSource } from './DataSource';
-import { RedisDataSourceOptions, RedisQuery } from './types';
+import { RedisQuery } from './redis';
+import { RedisDataSourceOptions } from './types';
 
 /**
  * Data Source plugin
