+++ src/redis/index.ts
@@ -1,1 +1,4 @@
 export * from './command';
+export * from './info';
+export * from './query';
+export * from './time-series';
