+++ src/types.ts
@@ -1,89 +1,4 @@
-import { DataQuery, DataSourceJsonData } from '@grafana/data';
-import { QueryTypeValue } from './redis';
-
-/**
- * Redis Query
- */
-export interface RedisQuery extends DataQuery {
-  /**
-   * Type
-   *
-   * @type {string}
-   */
-  type: QueryTypeValue;
-
-  /**
-   * Query command
-   *
-   * @type {string}
-   */
-  query?: string;
-
-  /**
-   * Field
-   *
-   * @type {string}
-   */
-  field?: string;
-
-  /**
-   * Redis TimeSeries filter
-   *
-   * @see https://oss.redislabs.com/redistimeseries/commands/#filtering
-   * @type {string}
-   */
-  filter?: string;
-
-  /**
-   * Command
-   *
-   * @type {string}
-   */
-  command?: string;
-
-  /**
-   * Key name
-   *
-   * @type {string}
-   */
-  key?: string;
-
-  /**
-   * Value label
-   *
-   * @type {string}
-   */
-  value?: string;
-
-  /**
-   * Aggregation
-   *
-   * @see https://oss.redislabs.com/redistimeseries/commands/#aggregation-compaction-downsampling
-   * @type {string}
-   */
-  aggregation?: string;
-
-  /**
-   * Bucket
-   *
-   * @type {string}
-   */
-  bucket?: string;
-
-  /**
-   * Legend label
-   *
-   * @type {string}
-   */
-  legend?: string;
-
-  /**
-   * Info Section
-   *
-   * @type {string}
-   */
-  section?: string;
-}
+import { DataSourceJsonData } from '@grafana/data';
 
 /**
  * Options configured for each DataSource instance
